refactor : 시간 표시 방법 변경 (24시 -> 12시)
